requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
build==1.2.1
    # via pip-tools
click==8.1.7
    # via pip-tools
django==5.0.4
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-filter==24.2
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.in
faker==24.11.0
    # via factory-boy
gunicorn==22.0.0
    # via -r requirements.in
jsonschema==4.21.1
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
packaging==24.0
    # via
    #   build
    #   gunicorn
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via faker
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via django
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
